add border

+++ app/src/main/java/com/example/retrofitrxjava/activity/HomeActivity.java
@@ -84,6 +84,7 @@ public class HomeActivity extends AppCompatAct<LayoutRecruitmentBinding> impleme
 
     @Override
     public void onItemDeleteClick(UserModel userModel) {
+        if (userModel.isAdmin()) return;
         appDatabase.getStudentDao().deleteAccount(userModel);
         adapterUser.setDt((ArrayList<UserModel>) appDatabase.getStudentDao().getAllUser());
     }
